Make a placed building appear, and answer for it when it cannot
The building's own message was the wrong shape. Its handle is two bytes, not
four, and four put two extra ones at the front of the branch, which shifted
everything behind it - a placed building went out complete and never appeared.
The stage is the ordinal the design's own rows are numbered by, not a count of
what is left to do, and the state block had its sale price eight bytes away from
where it is read, which is why a building came up owned by nobody.

Contributing to a building was impossible: interacting with one looked for a
doodad, then an NPC, then a mount, and never the building, so the player was
offered nothing to do with it. The stage's own skill is now offered, chosen by
the running total of actions the way the design partitions them.

Placement is checked before anything is taken. The plots of land and the rules
governing them were never read at all - a building could go up anywhere - so the
zone's plots, their groups, their opening dates and the categories they accept
now decide, and each refusal names its own reason instead of a generic one.
Turning a building costs what its design says it costs. A design's garden radius
moved to its own table in this version, so every building had a garden of zero.

Stages are looked up by the number the design gives them rather than by position
in the table, which finished a building a stage early whenever they did not start
at zero.

The area-configuration message was sending today's date, which only worked
because that moment had just passed: the client consults it before it will even
offer to place a building, and any clock skew would have had it refuse.

Also: the character list no longer sends a housing message this client does not
have - encoding it threw, so every account that owned a building lost its
character list.

+++ AAEmu.Game/Core/Packets/C2G/CSAllowHousingRecoverPacket.cs
@@ -12,7 +12,7 @@ public class CSAllowHousingRecoverPacket : GamePacket
 
     public override void Read(PacketStream stream)
     {
-        var tl = stream.ReadUInt16();
+        var tl = (ushort)stream.ReadUInt32(); // the handle is 32 bits on the wire; ours is 16
 
         Logger.Debug("AllowHousingRecover, Tl: {0}", tl);
         HousingManager.Instance.HousingToggleAllowRecover(Connection.ActiveChar, tl);
