+++ AAEmu.Game/Core/Packets/C2G/CSChangeHouseNamePacket.cs
@@ -12,7 +12,7 @@ public class CSChangeHouseNamePacket : GamePacket
 
     public override void Read(PacketStream stream)
     {
-        var tl = stream.ReadUInt16(); // houseId
+        var tl = (ushort)stream.ReadUInt32(); // the handle is 32 bits on the wire; ours is 16 // houseId
         var name = stream.ReadString();
 
         Logger.Debug("ChangeHouseName, Tl: {0}, Name: {1}", tl, name);
