+++ AAEmu.Game/Core/Packets/C2G/CSPerpayHouseTaxPacket.cs
@@ -11,7 +11,7 @@ namespace AAEmu.Game.Core.Packets.C2G
 
         public override void Read(PacketStream stream)
         {
-            var tl = stream.ReadUInt16();
+            var tl = (ushort)stream.ReadUInt32(); // the handle is 32 bits on the wire; ours is 16
             var ausp = stream.ReadBoolean();
 
             Logger.Debug("CSPerpayHouseTaxPacket, Tl: {0}, ausp: {1}", tl, ausp);
