+++ AAEmu.Game/Core/Packets/C2G/CSSellHousePacket.cs
@@ -13,8 +13,8 @@ public class CSSellHousePacket : GamePacket
 
     public override void Read(PacketStream stream)
     {
-        var tl = stream.ReadUInt16();
-        var moneyAmount = stream.ReadUInt32();
+        var tl = (ushort)stream.ReadUInt32(); // the handle is 32 bits on the wire; ours is 16
+        var moneyAmount = (uint)stream.ReadInt64(); // 64 bits on the wire
         var sellTo = stream.ReadString();
         var isPublic = stream.ReadBoolean();
         Logger.Debug($"SellHouse, Tl: {tl}, MoneyAmount: {moneyAmount}, SellTo: {sellTo}, isPublic: {isPublic}");
