+++ AAEmu.Game/Core/Packets/C2G/CSSetHouseAllowRecoverPacket.cs
@@ -11,7 +11,7 @@ namespace AAEmu.Game.Core.Packets.C2G
 
         public override void Read(PacketStream stream)
         {
-            var houseId = stream.ReadUInt16();  // tl
+            var houseId = (ushort)stream.ReadUInt32(); // the handle is 32 bits on the wire; ours is 16  // tl
             Logger.Debug("CSSetHouseAllowRecoverPacket, houseId: {0}", houseId);
         }
     }
